applyN functions moved to ApplyNFunctions trait
