+++ core/src/main/scala/cats/Bimonad.scala
@@ -2,4 +2,8 @@ package cats
 
 import simulacrum._
 
-@typeclass trait Bimonad[F[_]] extends Monad[F] with Comonad[F]
+trait Bimonad[F[_]] extends Monad[F] with Comonad[F]
+
+object Bimonad {
+  def apply[F[_]](implicit ev: Bimonad[F]): Bimonad[F] = ev
+}
