+++ core/src/main/scala/cats/FlatMap.scala
@@ -2,7 +2,7 @@ package cats
 
 import simulacrum._
 
-@typeclass trait FlatMap[F[_]] extends Apply[F] {
+trait FlatMap[F[_]] extends Apply[F] {
   def flatMap[A, B](fa: F[A])(f: A => F[B]): F[B]
 
   def flatten[A](ffa: F[F[A]]): F[A] =
@@ -10,4 +10,8 @@ import simulacrum._
 
   override def apply[A, B](fa: F[A])(ff: F[A => B]): F[B] =
     flatMap(ff)(f => map(fa)(f))
+}
+
+object FlatMap {
+  def apply[F[_]](implicit ev: FlatMap[F]): FlatMap[F] = ev
 }
