+++ core/src/main/scala/cats/Monad.scala
@@ -2,7 +2,11 @@ package cats
 
 import simulacrum._
 
-@typeclass trait Monad[F[_]] extends FlatMap[F] with Applicative[F] {
+trait Monad[F[_]] extends FlatMap[F] with Applicative[F] {
   override def map[A, B](fa: F[A])(f: A => B): F[B] =
     flatMap(fa)(a => pure(f(a)))
+}
+
+object Monad {
+  def apply[F[_]](implicit ev: Monad[F]): Monad[F] = ev
 }
